Add files via upload
The title of our project is “Speaking Hands: A sign language translation glove”. The working of our product is such that the gestures made by the disabled person will be sensed by gyroscope and flex sensors. The sensors will send the values which determine the positioning of hand and it will detect the gesture with the help of artificial intelligence. The meaning of the gesture will then be produced as a speech using voice module. This model solves a variety of problems from easing communication between mute and normal people, mute and blind people as well providing better understanding between disabled and normal people. The problem touches areas of AI/ML (data training), electronics (Arduino nano, flex sensors, and trans-receiver etc. ), physics (gyroscope and orientation) as well as mechanics and manufacturing.